softfonts: explicitly specify the default Pcss parameter (0) to DCS.
The mlterm terminal emulator supports DECDLD, but only accepts the DCS
command to declare the soft character if all the parameters are included
in the command. Previously we ommited the Pcss parameter (character set
size), leaving it to its default value (0=94 characters), which made
mlterm ignore the whole command. Adding the explicit 0 to the command,
makes mlterm happy.

This fix was suggested by James Holderness (github user j4james), in
github issue #6.